Move TUIO Input
